+++ R/sensitivity-calculation.R
@@ -5,7 +5,9 @@
 #' @param outputPaths Path (or a vector of paths) to the output(s) for which the
 #' sensitivity will be analyzed.
 #' @param parameterPaths A single or a vector of the parameter path(s) to be
-#' varied.
+#' varied. Can also be a named vector, where the names are user-defined labels.
+#' These names will be stored and used in downstream plotting functions (e.g.,
+#' as legend labels) if provided.
 #' @param variationRange Optional numeric vector or list defining the scaling of
 #' the parameters. The same variation range is applied to all specified parameters
 #' unless a list is provided, in which case the length of the list must match
@@ -90,6 +92,7 @@ sensitivityCalculation <- function(simulation,
   # Validate vector arguments of character type
   .validateCharVectors(outputPaths)
   .validateCharVectors(parameterPaths)
+  .validateCharVectors(names(parameterPaths), nullAllowed = TRUE)
   .validateCharVectors(variationType)
   .validateCharVectors(pkParameters, nullAllowed = TRUE)