+++ R/utilities-scenarios.R
@@ -127,7 +127,7 @@ runScenarios <- function(scenarios, simulationRunOptions = NULL) {
 #' simulated. See [createScenarios()] for details.
 #' @param customParams A list containing vectors 'paths' with the full paths to the
 #' parameters, 'values' the values of the parameters, and 'units' with the
-#' units the values are in. The values to be applied to the model.
+#' units the values are in. The values will be applied to all scenarios.
 #' @param stopIfParameterNotFound Boolean. If `TRUE` (default) and a custom parameter is not found, an error is thrown. If `FALSE`, non-existing parameters
 #' are ignored.
 #'