+++ tests/testthat/test-utilities.R
@@ -131,10 +131,6 @@ test_that("It finds a 0 with relative threshold", {
   expect_equal(getIndexClosestToValue(value = value, array = array, thresholdRel = 10), 2)
 })
 
-test_that("Escape a string for possible regular expression match", {
-  expect_equal(escapeForRegex("C:/Downloads"), "\\QC:/Downloads\\E")
-})
-
 test_that("`compareWithNA()` works as expected", {
   res <- compareWithNA(
     c(NA, "a", "b", NA),